requirements.txt
torch
torchvision
numpy
opencv-python
mediapipe
Pillow
# Optional: insightface (if using their pre-trained models directly)
# insightface
